requirements.txt
sounddevice
numpy
gradio
openai-whisper
